// dev-client-web-ignug-master/src/app/models/empresa/company.ts
export class Company {
    id: string;
    identity: string;
    nature: string;
    email: string;
    web_page: string;
    trade_name: string;
    comercial_activity: string;
    cell_phone: string;
    phone: string;
    address: string;
    avatar: string;
    state: string;

    constructor() {
      this.identity = '';
      this.nature = '';
      this.web_page = '';
      this.trade_name = '';
      this.comercial_activity = '';
      this.comercial_activity = '';
      this.cell_phone = '';
      this.phone = '';
      this.address = '';
      this.avatar = '';
      this.email = '';
    }
  }